Refactor operation helper to use generics for operator args, remove mapstructure
