+++ bot/divine.go
@@ -98,8 +98,5 @@ func _DivineCommand(message *discordgo.MessageCreate, args _DivineArgs) {
 		}
 	}
 
-	operationWrapper := func(srcBytes []byte, destBuffer io.Writer) error {
-		return Divine(srcBytes, destBuffer, args)
-	}
-	PrepareAndInvokeOperation(message, args.ImageURL, operationWrapper)
+	PrepareAndInvokeOperation(message, args.ImageURL, args, Divine)
 }
