+++ go.mod
@@ -1,12 +1,11 @@
 module github.com/fogo-sh/borik
 
-go 1.17
+go 1.18
 
 require (
 	github.com/bwmarrin/discordgo v0.22.0
 	github.com/joho/godotenv v1.3.0
 	github.com/kelseyhightower/envconfig v1.4.0
-	github.com/mitchellh/mapstructure v1.1.2
 	github.com/nint8835/parsley v0.0.0-20201224020611-dee14cbf9618
 	github.com/rs/zerolog v1.20.0
 	gopkg.in/gographics/imagick.v2 v2.6.0
